+++ packages/react-native-audio-api/common/cpp/utils/AudioUtils.h
@@ -1,25 +1,16 @@
+#pragma once
+
+#include <cstddef>
 
 namespace audioapi::AudioUtils {
-size_t timeToSampleFrame(double time, int sampleRate) {
-  return static_cast<size_t>(time * sampleRate);
-}
+size_t timeToSampleFrame(double time, int sampleRate);
 
-double sampleFrameToTime(int sampleFrame, int sampleRate) {
-  return static_cast<double>(sampleFrame) / sampleRate;
-}
+double sampleFrameToTime(int sampleFrame, int sampleRate);
 
 float linearInterpolate(
     const float *source,
     size_t firstIndex,
     size_t secondIndex,
-    float factor) {
-  if (firstIndex == secondIndex && firstIndex >= 1) {
-    return source[firstIndex] +
-        factor * (source[firstIndex] - source[firstIndex - 1]);
-  }
-
-  return source[firstIndex] +
-      factor * (source[secondIndex] - source[firstIndex]);
-}
+    float factor);
 
 } // namespace audioapi::AudioUtils
